Integrated existing movement functionality to movement package
